Rename API models
