+++ api/api/models/__init__.py
@@ -3,5 +3,5 @@
 # flake8: noqa
 from __future__ import absolute_import
 
-from api.models.basic_info import BasicInfo
-from api.models.security_token_response import TokenResponseModel
+from api.models.basic_info_model import BasicInfo
+from api.models.security_token_response_model import TokenResponseModel
